Add "synced" check to product publish

+++ facebook-commerce.php
@@ -828,17 +828,26 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 
 	/**
 	 * Generic function for use with any product publishing.
+	 *
 	 * Will determine product type (simple or variable) and delegate to
 	 * appropriate handler.
+	 *
+	 * @param int $wp_id product ID
 	 */
-	function on_product_publish( $wp_id ) {
+	public function on_product_publish( $wp_id ) {
+
 		if ( get_post_status( $wp_id ) != 'publish' ) {
 			return;
 		}
 
 		$woo_product  = new WC_Facebook_Product( $wp_id );
-		$product_type = $woo_product->get_type();
-		if ( WC_Facebookcommerce_Utils::is_variable_type( $woo_product->get_type() ) ) {
+
+		// skip if not enabled for sync
+		if ( ! $woo_product->woo_product instanceof \WC_Product || ! \SkyVerge\WooCommerce\Facebook\Products::product_should_be_synced( $woo_product->woo_product ) ) {
+			return;
+		}
+
+		if ( $woo_product->woo_product->is_type( 'variable' ) ) {
 			$this->on_variable_product_publish( $wp_id, $woo_product );
 		} else {
 			$this->on_simple_product_publish( $wp_id, $woo_product );
@@ -972,6 +981,11 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 
 		if ( ! $woo_product ) {
 			$woo_product = new WC_Facebook_Product( $wp_id, $parent_product );
+		}
+
+		// skip if not enabled for sync
+		if ( ! $woo_product->woo_product instanceof \WC_Product || ! \SkyVerge\WooCommerce\Facebook\Products::product_should_be_synced( $woo_product->woo_product ) ) {
+			return;
 		}
 
 		if ( $this->delete_on_out_of_stock( $wp_id, $woo_product ) ) {
@@ -2871,6 +2885,11 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 		$woo_product = new WC_Facebook_Product( $wp_id );
 		if ( ! $woo_product->exists() ) {
 			// This function can be called for non-woo products.
+			return;
+		}
+
+		// skip if not enabled for sync
+		if ( ! $woo_product->woo_product instanceof \WC_Product || ! \SkyVerge\WooCommerce\Facebook\Products::product_should_be_synced( $woo_product->woo_product ) ) {
 			return;
 		}
